Update Customer Repository     - Change parameter type     - Implement query function

+++ Assignment/Assignment.API/Infrastructure/EF/Repositories/ICustomerRepository.cs
@@ -5,8 +5,8 @@ namespace Assignment.API.Infrastructure.EF.Repositories {
 
     public interface ICustomerRepository {
 
-        Task<Customer> GetCustomerWithTransactionById(int id);
+        Task<Customer> GetCustomerWithTransactionById(long id);
         Task<Customer> GetCustomerWithTransactionByEmail(string email);
-        Task<Customer> GetCustomerWithTransactionByEmailAndId(string email, int id);
+        Task<Customer> GetCustomerWithTransactionByEmailAndId(string email, long id);
     }
 }
